modify : like -> favorite 라우트 네이밍 변경 및 추가 기능 구현
