+++ controllers/favoriteController.js
@@ -1,11 +1,11 @@
-const likeService = require('../services/likeService');
+const favoriteService = require('../services/favoriteService');
 
 const changeLike = async (req, res, next) => {
   try {
-    const user = req.userid[0].id;
+    const user = req.user;
     const estate = req.params.id;
 
-    const isLike = await likeService.changeLike(user, estate);
+    const isLike = await favoriteService.changeLike(user, estate);
     const result = isLike ? 'ADDED' : 'DELETED';
     return res.status(200).json({ message: `LIKE_IS_${result}` });
   } catch (err) {
@@ -15,9 +15,11 @@ const changeLike = async (req, res, next) => {
 
 const getLikeEstates = async (req, res, next) => {
   try {
-    const user = req.userid[0].id;
-    const likeEstates = await likeService.getLikeEstates(user);
-    return res.status(200).json({ messgae: 'LIKE_ESTATES_IS_RENDERED' });
+    const user = req.user[0];
+    const likeEstates = await favoriteService.getLikeEstates(user);
+    return res
+      .status(200)
+      .json({ messgae: 'LIKE_ESTATES_IS_RENDERED', liked: likeEstates });
   } catch (err) {
     next(err);
   }
