--- routes/likeRoute.js
@@ -1,9 +1,0 @@
-const express = require('express');
-const router = express.Router();
-const likeController = require('../controllers/likeController');
-const { usersValidateToken } = require('../middlewares/usersValidateToken');
-
-router.get('/', usersValidateToken, likeController.getLikeEstates);
-router.get('/:id', usersValidateToken, likeController.changeLike);
-
-module.exports = router;
